+++ affogato-core-ktx/src/main/java/com/parsuomash/affogato/core/ktx/datetime/Date.kt
@@ -146,6 +146,25 @@ fun String.toDateOrNull(pattern: String = "EEE MMM dd HH:mm:ss zzz yyyy"): Date?
  *
  * Example:
  * ```Kotlin
+ * 1659814200000.toDate().format("MM/dd/yyyy") // 08/07/2022
+ * ```
+ * @since 1.1.1
+ * @throws IllegalArgumentException if the given pattern is invalid
+ * @return The formatted date-time string.
+ * @see SimpleDateFormat
+ */
+fun Date.format(pattern: String): String {
+  simpleDateFormat.applyPattern(pattern)
+  return simpleDateFormat.format(this)
+}
+
+// TODO v2.0.0 remove deprecated functions ---------------------------------------------------------
+
+/**
+ * Formats a [Date] into a date-time [String].
+ *
+ * Example:
+ * ```Kotlin
  * 1659814200000.toDate().toString("MM/dd/yyyy") // 08/07/2022
  * ```
  * @since 1.1.0
@@ -157,29 +176,12 @@ fun String.toDateOrNull(pattern: String = "EEE MMM dd HH:mm:ss zzz yyyy"): Date?
   message = "This function is deprecated and will be removed in next major release." +
     "Use format() instead.",
   replaceWith = ReplaceWith(
-    expression = "format",
+    expression = "format(format)",
     imports = ["com.parsuomash.affogato.core.ktx.datetime.format"]
   ),
   level = DeprecationLevel.WARNING
 )
 fun Date.toString(format: String): String {
   simpleDateFormat.applyPattern(format)
-  return simpleDateFormat.format(this)
-}
-
-/**
- * Formats a [Date] into a date-time [String].
- *
- * Example:
- * ```Kotlin
- * 1659814200000.toDate().format("MM/dd/yyyy") // 08/07/2022
- * ```
- * @since 1.1.1
- * @throws IllegalArgumentException if the given pattern is invalid
- * @return The formatted date-time string.
- * @see SimpleDateFormat
- */
-fun Date.format(pattern: String): String {
-  simpleDateFormat.applyPattern(pattern)
   return simpleDateFormat.format(this)
 }