+++ affogato-core-ktx/src/main/java/com/parsuomash/affogato/core/ktx/datetime/LocalDate.kt
@@ -131,6 +131,25 @@ fun String.toLocalDateOrNull(pattern: String = "EEE MMM dd HH:mm:ss zzz yyyy"): 
  *
  * Example:
  * ```Kotlin
+ * 1659814200000.toLocalDate().format("MM/dd/yyyy") // 08/07/2022
+ * ```
+ * @since 1.1.1
+ * @throws IllegalArgumentException if the given pattern is invalid
+ * @return The formatted date-time string.
+ * @see SimpleDateFormat
+ */
+fun LocalDate.format(pattern: String): String {
+  simpleDateFormat.applyPattern(pattern)
+  return simpleDateFormat.format(toDate())
+}
+
+// TODO v2.0.0 remove deprecated functions ---------------------------------------------------------
+
+/**
+ * Formats a [LocalDate] into a date-time [String].
+ *
+ * Example:
+ * ```Kotlin
  * 1659814200000.toLocalDate().toString("MM/dd/yyyy") // 08/07/2022
  * ```
  * @since 1.1.0
@@ -142,29 +161,12 @@ fun String.toLocalDateOrNull(pattern: String = "EEE MMM dd HH:mm:ss zzz yyyy"): 
   message = "This function is deprecated and will be removed in next major release." +
     "Use format() instead.",
   replaceWith = ReplaceWith(
-    expression = "format",
+    expression = "format(format)",
     imports = ["com.parsuomash.affogato.core.ktx.datetime.format"]
   ),
   level = DeprecationLevel.WARNING
 )
 fun LocalDate.toString(format: String): String {
   simpleDateFormat.applyPattern(format)
-  return simpleDateFormat.format(toDate())
-}
-
-/**
- * Formats a [LocalDate] into a date-time [String].
- *
- * Example:
- * ```Kotlin
- * 1659814200000.toLocalDate().format("MM/dd/yyyy") // 08/07/2022
- * ```
- * @since 1.1.1
- * @throws IllegalArgumentException if the given pattern is invalid
- * @return The formatted date-time string.
- * @see SimpleDateFormat
- */
-fun LocalDate.format(pattern: String): String {
-  simpleDateFormat.applyPattern(pattern)
   return simpleDateFormat.format(toDate())
 }