+++ affogato-structure/src/main/java/com/parsuomash/affogato/structure/EntityMapper.kt
@@ -1,3 +1,5 @@
+@file:Suppress("unused")
+
 package com.parsuomash.affogato.structure
 
 /**