+++ affogato-structure/src/main/java/com/parsuomash/affogato/structure/validation/Validation.kt
@@ -1,0 +1,10 @@
+package com.parsuomash.affogato.structure.validation
+
+/**
+ * ## Validation
+ * Check is value valid.
+ * @since 1.4.0
+ */
+fun interface Validation {
+  fun isValid(): Boolean
+}